Increases the interval between retries (#505)
When looking through Sentry errors we thought we could maybe look into the retry intervals and duration, and see if there was anything to update.

Before we would wait 1 minute to start, I think we could increase it. So I copied the both the start and interval from the archiver_worker. We wait longer to start and between retries.

References: CV2-6037
PR 505