add rescue JSON::ParserError
My confusion was caused by the fact that sometimes archive org responds
with HTML instead of JSON, so we hit a JSON::ParserError. That would get
rescued and wouldn't hit the condition I expected.

That said, I think it makes sense to rescue JSON::ParserError. We can rescue
it, if it's more specific, e.g.: too many requests, we send that, if it's
not might be worth sending a generic JSON::ParserError.