6357 – Fix how we parse a generic Facebook watch page (#542)
A Facebook video from Facebook watch was sometimes redirecting to the main watch page. Meaning we were getting pretty generic content, like:

    URL: https://www.facebook.com/watch/
    Title: "Discover Popular Videos"

This caused `This item already exists` to be raised for different items. So, if a URL redirects to the main watch page, we want the title and URL to be the URL from the request.

To fix this issue, I added `/^https:\/\/([^\.]+\.)?facebook.com\/watch\/*?$/` to facebook_provider list of ignored URLs.

Regarding tests:
1. I broke 'sets fallbacks from metatags for event and watch URLS on apify error'
into two tests: one for events and one for watch URLs
– The reason: I wanted to make the expected behavior for watch URLs clearer:
     - If it redirects to an existing page:
          - we want to set title and description to what we can grab from the page
     - If it redirects to the main watch page:
          - we want to set title to nil and description to an empty string,
          - this way Media will set the title to the original URL,
          - description will be left empty.
          - We want to do this because otherwise we get the same title and
          description for a bunch of different links.

2. I also removed the two tests I had written before, so we only test
the `Parser::FacebookItem`, instead of testing Media as well. We expect
Media to set the fallbacks successfully, that is tested elsewhere.


References: 6357