change set_facebook_privacy_error to set_facebook_dead_end_error
This describes the method more accurately:
this method doesn't only set the privacy error, but it checks if a page
is unavailable, or, more accurately, a dead end, and then sets the error.

The reason why a page is a dead end isn't always because of a privacy_error
or login_required. Right now, we have two situations, consent page and
watch page, where we get redirected to dead ends for reasons other than
a privacy/login issue.