update to deal with video item response from apify
When we parsed videos, even though we got an answer from apify, we were
mostly using the data we could parse from the HTML. So this updates the
apify provider to also look for the video related keys.

I added a new response from apify so we could test this, and moved both
the new one and the old apify response to the data folder, so that the
test file is cleaner.

I also took the opportunity to refactor a bit. I thought there was a bit
of duplication.