6357 – Deal with video item response from Apify (#545)
When we parsed videos, we got an answer from Apify, but didn't use it. Because the keys in the Apify response are different between videos and posts. We were only looking for the posts' keys. I think we didn't notice this, because we were able to use the data we could parse from the HTML.

This updates the Apify provider to also look for the video related keys.

I also took the opportunity to refactor a bit. I thought there was a bit of duplication.

References: CV2-6357