update test to not make live request
This test started failing because requests to the nytimes require the proxy,
since that failed, the test failed as well.

But since we were already stubbing the ld_json return, I don't see why we
would need a live request at all. I think it is safe to use WebMock here.