update test to not make live request
This test was making a live request to the New York Times website,
which was failing due to a proxy issue. But, I don't think this needs
to be a live request, so I removed it from the page_item_test integration
tests and added it to the page_item_test unit tests.