Update failing tests to not make live requests (#557)
We had two failing tests caused by a failed request to the New York Times, it relies on a proxy to work, and sometimes it might just fail.

I looked into both tests and I think there is no need for them to be: live requests, specifically to the New York Times. So, I updated both to use stubs, and move one from the integration tests file to the models/units one.

- test "should parse page when json+ld tag content is an empty array"
But since we were already stubbing the ld_json return, I don't see why we
would need a live request at all. I think it is safe to use WebMock here.

- test "should store metatags in an Array"
I don't think this needs to be a live request, so I removed it from the page_item_test integration
tests and added it to the page_item_test unit tests.