Add test and new parse_nonmandatory_url method (#560)
If the author or author_picture urls are invalid, we should not stop the
item from being parsed.

I thought the best way to handle this was to create a new method
parse_nonmandatory_url in the RequestHelper module. This
way, we can safely attempt to parse URLs that might not be critical
to the media item.

It is only a wrapper around the existing parse_url method, so if there
are changes to parsing itself, we should only need to update
parse_url.

References: CV2-6414