handle_apify_error method should not fail when the parsed_response is a hash (#566)
Check if response is a hash or array and handle accordingly

Most of the time, it seems the apify response is an array with a hash,
but sometimes it can be a hash. Which was causing a TypeError. E.g.:

a = {a: 1}.first returns [:a, 1]
a[:a] returns 'no implicit conversion of Symbol into Integer (TypeError)'