update deploy action set-output command (#570)
This deprecation seems to be planned for a while now, this updates to use the new syntax so we get the rid of the warning:

'The set-output command is deprecated and will be disabled soon. Please upgrade to using Environment Files. For more information see: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/'