6210 – Remove OpenUriRedirections gem (#571)
We were using the OpenUriRedirections gem to fix an issue that no longer exists in Open URI: HTTP to HTTPS is supported by native open-uri by default from Ruby 2.4 (https://github.com/open-uri-redirections/open_uri_redirections).

So we are removing it and using Open URI instead: The one we were using was only intended as a patch, and hasn't been maintained in a long time; Open URI is a well maintained internal ruby library. We also think it might help with a redirection loop error we have been seeing with all Instagram links.

We had a CodeQl security alert in this ticket regarding the usage of URI.open, but we dismissed it: The uri being passed is not a string, it has been parsed, and is already an Addressable URI object. Since we are using Addressable, we can't re-write this the way CodeQL suggests. But, any malicious code would not get to this point, due to uri being an Addressable URI object. 

References: CV2-6210