6210 – skip_get_canonical_url when the host is Instagram (#572)
When we try to get the canonical URL for an Instagram media in production
environment, it always results in a redirection loop. This error happens
in OpenURI, so there is not a lot we can do to fix it.

The error does not affect the functionality of the application,
but it causes unnecessary processing and is really noisy (since we get a
notification in Sentry every time).

Since the parsing of instagram links works fine without the canonical URL,
we are skipping this step, but logging when we do, if we need to review
at some point.

References: CV2-6210