Rename as_json to process_and_return_json
The intent behind the name change is to signal that this method does not
olny return a json object, but orchestrates multiple steps:
parsing, caching, archiving, uploading images, and returning data.

Should this all be happening here, like this? Maybe not, but this is what
is currently happening, so maybe a first step is to make that clearer.