Skip get_canonical_url for Twitter items
The OpenURI request fails with a 400, because there is authorization that
is specific to the Twitter parser, that Media does not have access to.

So since we know this will fail, we skip it. But it is another sign
that this is Parser logic inside of Media, and it needs to be refactored.