Update minitest to 5.25.5 (#587)
I have been having pretty consistent issues with hanging tests, after some research it turned up that maybe this could be an issue with a version of Minitest.

I updated it locally, and it didn't hang, but sometimes it seems fine for a while, only for the issue to still be present.

Hopefully this fixes it.

References: CV2-6431