Clean up fallback logic
I added a DATA_STRUCTURE constant to define the expected structure of
the data hash. This helps as documentation, but we can also initialize
with a deep copy of this structure to ensure all keys are present, because
some methods rely on the presence of certain keys.

I also think fallback and required_fields should be instance methods,
and we shouldn't need a separate minimal_data method anymore.
But we do need it for get_error_data, since I haven't found a better
solution yet, I kept this as a middle ground.