Add more info to logs when webhook notification fails with 4xx/5xx error (#601)
Since we implemented this fix (https://github.com/meedan/check-api/pull/2345) in Check-API, we are seeing way less of this. But we still see a few, specially related to Perma.cc.

So I'm adding some more information to the logs to try to get more info, and then hopefully deal with this in a better way.

References: CV2-5395