Fix get_metadata_from_tag to handle nil doc
I had been seeing this error in Sentry, but originally mistook it for
something else. It was a small fix, just need to return early if doc is nil.

I added a couple of tests to ensure that no error is raised when doc is nil,
both at this method level and at the parse_data level.